phase5_chat_interface/backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
sentence-transformers
numpy
groq
